+++ BraintreeDropIn/BTCardFormViewController.m
@@ -21,6 +21,11 @@
 #else
 #import <BraintreeUnionPay/BraintreeUnionPay.h>
 #endif
+#if __has_include("BraintreePaymentFlow.h")
+#import "BraintreePaymentFlow.h"
+#else
+#import <BraintreePaymentFlow/BraintreePaymentFlow.h>
+#endif
 
 @interface BTCardFormViewController () <BTViewControllerPresentingDelegate>
 
@@ -179,6 +184,7 @@
     self.expirationDateField.delegate = self;
     self.securityCodeField = [[BTUIKSecurityCodeFormField alloc] init];
     self.securityCodeField.delegate = self;
+    self.securityCodeField.textField.secureTextEntry = self.dropInRequest.shouldMaskSecurityCode;
     self.postalCodeField = [[BTUIKPostalCodeFormField alloc] init];
     self.postalCodeField.delegate = self;
     self.mobileCountryCodeField = [[BTUIKMobileCountryCodeFormField alloc] init];
@@ -592,6 +598,7 @@
     UIBarButtonItem *addCardButton = self.navigationItem.rightBarButtonItem;
     self.navigationItem.rightBarButtonItem = [[UIBarButtonItem alloc] initWithCustomView:spinner];
     self.view.userInteractionEnabled = NO;
+    __block UINavigationController *navController = self.navigationController;
 
     [cardClient tokenizeCard:cardRequest options:nil completion:^(BTCardNonce * _Nullable tokenizedCard, NSError * _Nullable error) {
         dispatch_async(dispatch_get_main_queue(), ^{
@@ -600,23 +607,37 @@
             self.navigationItem.rightBarButtonItem = addCardButton;
 
             if (self.dropInRequest.threeDSecureVerification && self.dropInRequest.amount != nil
-                && [self.configuration.json[@"threeDSecureEnabled"] isTrue] && [[BTTokenizationService sharedService] isTypeAvailable:@"ThreeDSecure"]) {
+                && [self.configuration.json[@"threeDSecureEnabled"] isTrue]) {
 
-                NSMutableDictionary *options = [[NSMutableDictionary alloc] init];
-                options[BTTokenizationServiceViewPresentingDelegateOption] = self;
-                options[BTTokenizationServiceAmountOption] = [[NSDecimalNumber alloc] initWithString:self.dropInRequest.amount];
-                options[BTTokenizationServiceNonceOption] = tokenizedCard.nonce;
+                BTPaymentFlowDriver *paymentFlowDriver = [[BTPaymentFlowDriver alloc] initWithAPIClient:self.apiClient];
+                paymentFlowDriver.viewControllerPresentingDelegate = self;
 
-                [[BTTokenizationService sharedService] tokenizeType:@"ThreeDSecure" options:options withAPIClient:self.apiClient completion:^(BTPaymentMethodNonce * _Nullable tokenizedCard, NSError * _Nullable error) {
-                    if (tokenizedCard || error) {
-                        [self.delegate cardTokenizationCompleted:tokenizedCard error:error sender:self];
-                    } else {
-                        [self cancelTapped];
+                BTThreeDSecureRequest *request = [[BTThreeDSecureRequest alloc] init];
+                request.amount = [[NSDecimalNumber alloc] initWithString:self.dropInRequest.amount];
+                request.nonce = tokenizedCard.nonce;
+                [paymentFlowDriver startPaymentFlow:request completion:^(BTPaymentFlowResult * _Nonnull result, NSError * _Nonnull error) {
+                    if (error) {
+                        if (error.code == BTPaymentFlowDriverErrorTypeCanceled) {
+                            [self cancelTapped];
+                        } else {
+                            [self.delegate cardTokenizationCompleted:nil error:error sender:self];
+                        }
+                    } else if (result) {
+                        BTThreeDSecureResult *threeDSecureResult = (BTThreeDSecureResult *)result;
+                        [self.delegate cardTokenizationCompleted:threeDSecureResult.tokenizedCard error:error sender:self];
                     }
                 }];
-
             } else {
-                [self.delegate cardTokenizationCompleted:tokenizedCard error:error sender:self];
+                if (error != nil) {
+                    dispatch_async(dispatch_get_main_queue(), ^{
+                        UIAlertController *alertController = [UIAlertController alertControllerWithTitle:BTUIKLocalizedString(CARD_DETAILS_LABEL) message:BTUIKLocalizedString(REVIEW_AND_TRY_AGAIN) preferredStyle:UIAlertControllerStyleAlert];
+                        UIAlertAction *alertAction = [UIAlertAction actionWithTitle:BTUIKLocalizedString(TOP_LEVEL_ERROR_ALERT_VIEW_OK_BUTTON_TEXT) style:UIAlertActionStyleDefault handler:nil];
+                        [alertController addAction: alertAction];
+                        [navController presentViewController:alertController animated:YES completion:nil];
+                    });
+                } else {
+                    [self.delegate cardTokenizationCompleted:tokenizedCard error:error sender:self];
+                }
             }
         });
     }];