+++ BraintreeDropIn/Public/BTDropInRequest.h
@@ -42,13 +42,17 @@ NS_ASSUME_NONNULL_BEGIN
 /// Defaults to false. Set to true to hide the Venmo option even if enabled for your account.
 @property (nonatomic, assign) BOOL venmoDisabled;
 
-/// Optional: If true and an amount is set, ThreeDSecure will be used to verify the card. ThreeDSecure must be enabled in the control panel.
+/// Optional: If true and an amount is set, ThreeDSecure will be used to verify new cards. ThreeDSecure must be enabled in the control panel.
 /// Defaults to false.
 @property (nonatomic, assign) BOOL threeDSecureVerification;
 
 /// Controls if payment methods should be tokenized and stored in the user's Braintree vault. Set to false when doing transactions externally (i.e. in Safari browser - which does the vaulting), otherwise set to true
 /// Default to false.
 @property (nonatomic, assign) BOOL shouldTokenize;
+
+/// Optional: If true the security code will be masked.
+/// Defaults to false.
+@property (nonatomic, assign) BOOL shouldMaskSecurityCode;
 
 @end